Proper Implementation of Transposition Tables
Before:
Running MTDf and NegaMax return different bestmoves for the same depth.
Example is depth 4 for the starting position. We only want to compare the zobrist key of the stored entry with the zobrist. Since we already check if key is in the TT we dont actually need to compare our key with the saved key. This is because of pythons dictionary use. If one were to use lists instead of an dictionary this step is necessary.
Notice that NegaMax is actually faster than MTDf. I dont know why yet but I suspect the evaluation fluctuates too much?